server/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
django-filter>=24.0,<25.0
drf-spectacular>=0.28,<1.0
Pillow>=11.0,<12.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
